chore: resolve eslint errors
