+++ src/hooks/shared.hooks.ts
@@ -25,7 +25,7 @@ export const useMatrixClient = () => {
     };
 
     initClient();
-  }, []);
+  }, [matrixClient]);
 
   return matrixClient;
 };
